Moved EntryEditor to a separate directory
